+++ src/components/QueuePanel.tsx
@@ -1,6 +1,5 @@
 import { getPlaylist, updatePlaylist } from '../api/subsonicPlaylists';
 import { buildCoverArtUrl, coverArtCacheKey } from '../api/subsonicStreamUrl';
-import { getAlbum } from '../api/subsonicLibrary';
 import { registerQueueListScrollTopReader, consumePendingQueueListScrollTop } from '../store/queueUndo';
 import { songToTrack } from '../utils/songToTrack';
 import type { Track } from '../store/playerStoreTypes';
@@ -21,7 +20,6 @@ import { copyTextToClipboard } from '../utils/serverMagicString';
 import { showToast } from '../utils/toast';
 import { useThemeStore } from '../store/themeStore';
 import { useLyricsStore } from '../store/lyricsStore';
-import { useDragDrop, registerQueueDragHitTest } from '../contexts/DragDropContext';
 import LyricsPane from './LyricsPane';
 import NowPlayingInfo from './NowPlayingInfo';
 import { TFunction } from 'i18next';
@@ -36,6 +34,7 @@ import { SavePlaylistModal } from './queuePanel/SavePlaylistModal';
 import { LoadPlaylistModal } from './queuePanel/LoadPlaylistModal';
 import { QueueHeader } from './queuePanel/QueueHeader';
 import { QueueCurrentTrack } from './queuePanel/QueueCurrentTrack';
+import { useQueuePanelDrag } from '../hooks/useQueuePanelDrag';
 
 export default function QueuePanel() {
   const orbitRole = useOrbitStore(s => s.role);
@@ -216,9 +215,6 @@ function QueuePanelHostOrSolo() {
     if (!expandReplayGain) setLufsTgtOpen(false);
   }, [expandReplayGain]);
 
-  // Tracks which queue index is being psy-dragged for opacity visual feedback
-  const psyDragFromIdxRef = useRef<number | null>(null);
-
   const queueListRef = useRef<HTMLDivElement>(null);
 
   useLayoutEffect(() => {
@@ -238,113 +234,20 @@ function QueuePanelHostOrSolo() {
 
   const asideRef = useRef<HTMLElement>(null);
 
-  useEffect(() => {
-    const hitTest = (cx: number, cy: number) => {
-      const el = asideRef.current;
-      if (!el) return false;
-      const r = el.getBoundingClientRect();
-      return cx >= r.left && cx <= r.right && cy >= r.top && cy <= r.bottom;
-    };
-    return registerQueueDragHitTest(hitTest);
-  }, []);
-
-  const { isDragging: isPsyDragging, startDrag, payload: psyPayload } = useDragDrop();
-  /** Only these drag types may be dropped into the queue. */
-  const QUEUE_DROP_TYPES = new Set(['song', 'album', 'queue_reorder']);
-  const isQueueDrag = isPsyDragging && !!psyPayload && (() => {
-    try { return QUEUE_DROP_TYPES.has(JSON.parse(psyPayload.data).type); } catch { return false; }
-  })();
-  // Keep for the onPsyDrop radio-reject check below
-  const isRadioDrag = isPsyDragging && !!psyPayload && (() => {
-    try { return JSON.parse(psyPayload.data).type === 'radio'; } catch { return false; }
-  })();
-
-  useEffect(() => {
-    if (!isPsyDragging) {
-      externalDropTargetRef.current = null;
-      setExternalDropTarget(null);
-    }
-  }, [isPsyDragging]);
-
-  const [externalDropTarget, setExternalDropTarget] = useState<{ idx: number; before: boolean } | null>(null);
-  const externalDropTargetRef = useRef<{ idx: number; before: boolean } | null>(null);
-
-  // ── Mouse-event DnD: listen for psy-drop custom events ─────────
-  useEffect(() => {
-    const aside = asideRef.current;
-    if (!aside) return;
-
-    const onPsyDrop = async (e: Event) => {
-      const detail = (e as CustomEvent).detail;
-      if (!detail?.data) return;
-
-      let parsedData: any = null;
-      try { parsedData = JSON.parse(detail.data); } catch { return; }
-
-      // Radio streams are not tracks — reject silently
-      if (parsedData.type === 'radio') return;
-
-      const dropTarget = externalDropTargetRef.current;
-      externalDropTargetRef.current = null;
-      setExternalDropTarget(null);
-
-      const insertIdx = dropTarget
-        ? (dropTarget.before ? dropTarget.idx : dropTarget.idx + 1)
-        : usePlayerStore.getState().queue.length;
-
-      if (parsedData.type === 'queue_reorder') {
-        const fromIdx: number = parsedData.index;
-        psyDragFromIdxRef.current = null;
-        if (fromIdx !== insertIdx) reorderQueue(fromIdx, insertIdx);
-      } else if (parsedData.type === 'song') {
-        enqueueAt([parsedData.track], insertIdx);
-      } else if (parsedData.type === 'songs') {
-        enqueueAt(parsedData.tracks as Track[], insertIdx);
-      } else if (parsedData.type === 'album') {
-        const albumData = await getAlbum(parsedData.id);
-        const tracks: Track[] = albumData.songs.map((s: any) => ({
-          id: s.id, title: s.title, artist: s.artist, album: s.album,
-          albumId: s.albumId, artistId: s.artistId, duration: s.duration, coverArt: s.coverArt, track: s.track,
-          year: s.year, bitRate: s.bitRate, suffix: s.suffix, userRating: s.userRating, genre: s.genre,
-        }));
-        enqueueAt(tracks, insertIdx);
-      }
-    };
-
-    aside.addEventListener('psy-drop', onPsyDrop);
-    return () => aside.removeEventListener('psy-drop', onPsyDrop);
-  }, [enqueueAt]);
-
-  // Drag a queue row outside the panel → remove (drop never reaches `aside`).
-  useEffect(() => {
-    const onDocPsyDrop = (e: Event) => {
-      if (!isQueueVisible) return;
-      const d = (e as CustomEvent<{ data?: string; clientX?: number; clientY?: number }>).detail;
-      if (!d?.data) return;
-      const cx = d.clientX;
-      const cy = d.clientY;
-      if (typeof cx !== 'number' || typeof cy !== 'number') return;
-      let parsed: { type?: string; index?: number } | null = null;
-      try {
-        parsed = JSON.parse(d.data);
-      } catch {
-        return;
-      }
-      if (parsed?.type !== 'queue_reorder' || typeof parsed.index !== 'number') return;
-      const aside = asideRef.current;
-      if (!aside) return;
-      const r = aside.getBoundingClientRect();
-      const inside =
-        cx >= r.left && cx <= r.right && cy >= r.top && cy <= r.bottom;
-      if (inside) return;
-      psyDragFromIdxRef.current = null;
-      externalDropTargetRef.current = null;
-      setExternalDropTarget(null);
-      removeTrack(parsed.index);
-    };
-    document.addEventListener('psy-drop', onDocPsyDrop);
-    return () => document.removeEventListener('psy-drop', onDocPsyDrop);
-  }, [isQueueVisible, removeTrack]);
+  const {
+    psyDragFromIdxRef,
+    externalDropTarget,
+    externalDropTargetRef,
+    setExternalDropTarget,
+    isQueueDrag,
+    startDrag,
+  } = useQueuePanelDrag({
+    asideRef,
+    isQueueVisible,
+    reorderQueue,
+    enqueueAt,
+    removeTrack,
+  });
 
   useEffect(function queueAutoScroll() {
     if (suppressNextAutoScrollRef.current) {
